add market data to Buy flow
