+++ src/i18n/offerPreferences/en.json
@@ -16,6 +16,8 @@
   "offerPreferences.buyOffers" : "$0 buy offers",
   "offerPreferences.competingSellOffers" : "$0 competing sell offers",
   "offerPreferences.competingSellOffersBelowThisPremium" : "$0 competing sell offers below this premium",
+  "offerPreferences.competingBuyOffers" : "$0 competing buy offers",
+  "offerPreferences.competingBuyOffersBelowThisPremium" : "$0 competing buy offers below this premium",
   "offerPreferences.currentPrice" : "currently $0",
   "offerPreferences.feature.instantTrade" : "instant - trade",
   "offerPreferences.filters" : "filters",
