+++ src/views/offerPreferences/CreateBuyOffer.tsx
@@ -35,6 +35,8 @@ import { MarketInfo } from "./components/MarketInfo";
 import { PreferenceMethods } from "./components/PreferenceMethods";
 import { PreferenceScreen } from "./components/PreferenceScreen";
 import { Section } from "./components/Section";
+import { useFilteredMarketStats } from "./components/useFilteredMarketStats";
+import { usePastOffersStats } from "./usePastOffersStats";
 import { usePostBuyOffer } from "./utils/usePostBuyOffer";
 import { useRestrictSatsAmount } from "./utils/useRestrictSatsAmount";
 import { useTradingAmountLimits } from "./utils/useTradingAmountLimits";
@@ -250,16 +252,30 @@ function PublishOfferButton() {
   );
 }
 
-function CompetingOfferStats() {
-  const text = tw`text-center text-success-dark-2`;
+export function CompetingOfferStats() {
+  const text = tw`text-center text-success-dark-2 subtitle-2`;
+
+  const meansOfPayment = useOfferPreferences((state) => state.meansOfPayment);
+  console.log("means of payment: ", meansOfPayment);
+  const { data: pastOfferData } = usePastOffersStats({ meansOfPayment });
+  const { data: marketStats } = useFilteredMarketStats({
+    type: "bid",
+    meansOfPayment,
+  });
 
   return (
     <Section.Container style={tw`gap-1 py-0`}>
       <PeachText style={text}>
-        {i18n("offerPreferences.competingSellOffers", "x")}
+        {i18n(
+          "offerPreferences.competingBuyOffers",
+          String(marketStats.offersWithinRange.length),
+        )}
       </PeachText>
       <PeachText style={text}>
-        {i18n("offerPreferences.premiumOfCompletedTrades", "9")}
+        {i18n(
+          "offerPreferences.premiumOfCompletedTrades",
+          String(pastOfferData?.avgPremium),
+        )}
       </PeachText>
     </Section.Container>
   );
